# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "APIs for Musicplayer app built as part of onboarding",
    "version": "1.0.0",
    "title": "Musicplayer Server"
  },
  "tags": [
    {
      "name": "playlist",
      "description": "Everything about your Pets"
    },
    {
      "name": "song",
      "description": "Access to Petstore orders"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/playlist": {
      "get": {
        "tags": [
          "playlist"
        ],
        "description": "Get info about playlist",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Playlist"
            }
          }
        }
      }
    },
    "/playlist/song": {
      "post": {
        "tags": [
          "playlist"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Song info",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SongInfo"
            }
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Song"
            }
          },
          "400": {
            "description": "Invalid input, input validation error"
          }
        }
      }
    },
    "/playlist/song/{id}": {
      "delete": {
        "tags": [
          "playlist"
        ],
        "description": "Remove song from playlist",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Song"
            }
          },
          "400": {
            "description": "Invalid Song ID"
          }
        }
      }
    },
    "/song/{id}": {
      "get": {
        "tags": [
          "song"
        ],
        "description": "Get info about playlist",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Song"
            }
          }
        }
      }
    },
    "/song/{id}/heart": {
      "post": {
        "tags": [
          "song"
        ],
        "description": "Update heart status of song",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "integer",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "description": "heart status",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "heartStatus"
              ],
              "properties": {
                "heartStatus": {
                  "type": "boolean"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Song"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Playlist": {
      "type": "object",
      "properties": {
        "lastUpdated": {
          "type": "integer",
          "format": "date-time"
        },
        "songs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SongMeta"
          }
        }
      }
    },
    "SongMeta": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "title": {
          "type": "string"
        }
      }
    },
    "Song": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "title": {
          "type": "string"
        },
        "artist": {
          "type": "string"
        },
        "cover": {
          "type": "string",
          "format": "uri"
        },
        "url": {
          "type": "string",
          "format": "uri"
        },
        "heart": {
          "type": "boolean"
        }
      }
    },
    "SongInfo": {
      "type": "object",
      "required": [
        "title",
        "artist",
        "url"
      ],
      "properties": {
        "title": {
          "type": "string"
        },
        "artist": {
          "type": "string"
        },
        "cover": {
          "type": "string",
          "format": "uri"
        },
        "url": {
          "type": "string",
          "format": "uri"
        },
        "heart": {
          "type": "boolean"
        }
      }
    }
  }
}